Remove unused files, move SVOs to assets folder
