+++ src/engine/Constants.java
@@ -7,6 +7,8 @@ public class Constants {
     public static final float CAMERA_LOWER_LIMIT = -1.570f;
     public static final float CAMERA_UPPER_LIMIT = 1.570f;
     public static final int OCTREE_MEMORY_SIZE_KB = 1000000;
+    public static final int CHUNK_SIZE = 1024;
+    public static final String MAP_DIR = "./assets/";
     public static final int[][] CHILD_OFFSETS = {
         {0, 0, 0},
         {1, 0, 0},
