+++ src/engine/OctreeThread.java
@@ -2,14 +2,14 @@ import java.nio.ByteBuffer;
 
 public class OctreeThread extends Thread{
 
-  EfficientOctree octree;
+  Octree octree;
   int[] parentPos;
   ByteBuffer voxelBuffer;
 
   public OctreeThread(int[] parentPos, ByteBuffer voxelBuffer){
     this.parentPos = parentPos;
     this.voxelBuffer = voxelBuffer;
-    octree = new EfficientOctree(Constants.OCTREE_MEMORY_SIZE_KB / 8, 512, parentPos);
+    octree = new Octree(Constants.OCTREE_MEMORY_SIZE_KB / 8);
   }
 
   @Override
